requirements.txt
Flask==0.10.1
foreman==0.9.7
futures==3.0.3
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyHyphen==2.0.5
requests==2.8.0
requests-futures==0.9.5
Werkzeug==0.10.4
wheel==0.24.0
